Invalid balances exception moved to chapter source
