--- Tests/BootCamp.Chapter.Tests/InvalidBalancesException.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace BootCamp.Chapter.Tests
-{
-    internal class InvalidBalancesException: Exception
-    {
-    }
-}
